We have fixed some bug in different section of our simulator (learning rules, integration method, neuron models, etc.). Additionally, we have included the following changes in our simulator:
-OpenMP implementation:
We have included new OpenMP support in EDLUT.  Now we can create a variable number of threads to process our neural network. This number of threads is defined with two new input parameter (-openmpQ and -openmp). EDLUT uses a new feature named "task" in OpenMP. This option is only available in 3.0 versions or higher (EDLUT check automatically the OpenMP version). This feature allows EDLUT to split "big tasks", such as the post-synaptic learning rule update or the time driven neuron model integration phase, in "small tasks" that can be performed by whatever thread. With this option we balance the computational load between the different threads.
-openmpQ  X, where X is the number of sections in which the neural network will be divided. EDLUT calculates the minimum inter-propagation time between the different neural network sections an creates a new event named "SynchronizeActivityEvent" that propagate the spikes between the sections.
-openmp Y, where Y>=X is the number of OpenMP threads that will be used. If the OpenMP version used is 3.0 or higher, the total number of threads will be Y. Otherwise, the total number of thread will be X. This additional number of threads can be used to perform "small OpenMP tasks".



-TableBasedModelHF:
When an input spike arrives to an event-driven neuron, three steps must be done:
1º) Update the neuron model variables.
2º) Increment the correspondent conductance value.
3º) Predict if the neuron will spike and when. In case of spike, this one must be inserted in the event queue.
If a second spike arrive to the same neuron at the same instant, only the second and third step are done. The problem it is that the prediction makes in the third step for the first spike it is no longer valid and the computation time consumed in this step was a waste of time. For this reason we have created a modified version of the original event-driven neuron model (TableBasedModel). This modified model reaches better performance levels when the input activity to its neuron is synchronous, but worst performance levels otherwise. This version is better than the original when the number of input synapses is pretty high, such as in the Purkinge layer.
In this case, when  an input spike arrives to a neuron in a time X, the following steps must be done:
1º) Update the neuron model until instant X.
2º) Increment the correspondent conductance value.
3º) Create a new event named "TableBasedModelHFEvent" for a time X+1us and insert this new event in the event queue.
If a new spike arrive to this neuron in this period of 1us, only the second step must be done. Finally, the TableBasedModelHFEvent predicts if the neuron will spike and when. In case of spike, this one is inserted in the event queue.



-Real time:
We have included a new real time option in order to operate properly with a real robot. In this case, EDLUT must perform chunk of 2ms of simulation in less than 2ms of execution time. When EDLUT real time simulation feature is enabled, a processing thread is then created playing the role of a "watchdog timer". This thread is responsible for checking the simulation time consumption per time slot and holds the needed simulating variables that ensure real time (see Table 1 and 2 below). The main time slot that ensures real time is divided into four micro-step times whose time durations are normalized. Each micro-step time has assigned a certain restriction level ranged from zero to three depending on how close to the real time boundary the simulation is running. Restriction level “Zero” is activated by default. The simulation remains under this restriction level while the simulation time does not surpass the first micro-step timing. Restriction level “One” is activated in case the simulation time does surpass the first micro-step timing but still remains below the second micro-step timing. When the second micro-step timing is indeed surpassed the restriction level “Two” is activated. The more critical restriction level (level 3) occurs when the simulation time is about the same magnitude than the last micro-step into which the main time slot has been divided, that is, the simulation is running close to the time-step boundary that ensures real time.
Table 1 Restriction levels when EDLUT real time feature is enabled

Restriction level 		Interval 
0				[0, 1st micro-step time] 
1 				[1st micro-step time, 2nd micro-step time] 
2 				[2nd micro-step time,3rd  micro-step time]  
3				[3rd  micro-step time, main time slot]

Depending on the level of restriction that the simulation is running, different contingency actions within EDLUT simulator are carried out:

Table 2 Contingency tasks related to restriction levels when EDLUT real time feature is enabled.
Restriction level 		Contingency tasks 	
0 				Standard simulation. 
1 				Learning laws are disabled. 
2 				Both spike processing and neuronal model updates are disabled.
3 				All EDLUT processes in actions are disabled.



-Learning rules:
We include the new learning rule files described in previous versions (we forgot to upload these files the last time).