- Bug fixed in configure file for Linux and Mac: CUDA option works now properly in both operative systems.
- Variable step integration methods has been discarded in this new version.

- Computation activity (spikes) has been optimized: Several modifications have be done in order to improve the spike management. The propagation delay of each synapse 
must be now a multiple of 0.0001 seconds.

- Multiples GPUs: EDLUT checks the number of CUDA capable GPUs and uses all the available ones.

- New time-driven neuron model in CPU and GPU (TimeDrivenPurkinjeCell and TimeDrivenPurkinjeCell_GPU): This new neuron model represents a Purkinge Cell. Its configuration file only must containe the integration method used (Euler, RK2, RK4, RK45, BDF1, BDF2, ..., BDF6) and the integration step in seconds.

- Random generator for linux and windows: The rand() function defined in math.h has different implementations for Windows and Linux. For this reason we have implemented our own random number generator.

- Bug fixed in the definition of two learning rules in a single synapse.

- Some time driven neuron models (LIFTimeDrivenModel_1_4, EgidioGranuleCell_TimeDriven and TimeDrivenPurkinjeCell) have been optimized pre-computing some equations in static tables.

- WE ARE WORKING NOW IN A BRIEF TUTORIAL FOR EDLUT. IT WILL BE INCLUDED IN THIS REPOSITORY AS SOON AS POSIBLE. WE WILL DEFINE IN THIS DOCUMENT THE MAIN OPTIONS OF EDLUT, INCLUDING THE FOLLOWING MODIFICACION:

A) New executable file (RealTimeEDLUTKernel): This one works as the executable EDLUTKernel. The main difference it is that it can be executed in real time. Parameter -rt, -rtgap, -rt1, -rt2 and -rt3 must be used.


B) NEW WAY OF DEFINE SUPERVISED LEARNING RULES IN NETWORK FILE: Each learning rule is now defined in a single line. In the network file, when the synapses are defined, the last parameter indicates the associated learning rule. If this learning rule is supervised, we must indicate which synapses are "normal" and which ones are "trigger". When we want to specify a trigger connection, we include the letter "t" before the learning rule number.

0 1 1 1 1 0.001 0 0 10.0000 0
0 1 1 1 1 0.001 0 0 10.0000 t0


C) New learning rules have been included: SimetricCosSinAdditiveKernel, SimetricCosSinSTDPAdditiveKernel and SimetricCosSTDPAdditiveKernel.